# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Epublib - OpenAPI 3.0
  description: |-
    Swagger 3.0 Documentation of Chemons APIs Endpoint
  contact:
    email: fuad.mustamirrul@gmail.com
  version: 1.0.11
...
paths:
...
  /api/v1/reset-password/request:
    post:
      tags:
        - Reset Password
      summary: Request Reset Password Email
      operationId: resetPasswordRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                status: 200
                message: "Success"
                data: {}
        '400':
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
              example:
                status: 400
                message: "Invalid email format"
                data: {}
...
components:
  schemas:
    GenericResponse:
      required:
        - status
        - message
      type: object
      properties:
        status: 
          type: integer
          example: 200
        message: 
          type: string
          example: Success
        data: 
          type: object
...
    ResetPasswordRequest:
      type: object
      properties:
        email:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer